+++ src/main/java/com/atlauncher/gui/dialogs/InstanceExportDialog.java
@@ -54,6 +54,7 @@ import com.atlauncher.managers.AccountManager;
 import com.atlauncher.utils.ComboItem;
 import com.atlauncher.utils.OS;
 import com.atlauncher.utils.Utils;
+import com.atlauncher.utils.WindowUtils;
 
 @SuppressWarnings("serial")
 public class InstanceExportDialog extends JDialog {
@@ -80,12 +81,12 @@ public class InstanceExportDialog extends JDialog {
             }
         });
 
+        WindowUtils.resizeForContent(this);
+
         setVisible(true);
     }
 
     private void setupComponents() {
-        setSize(550, 460);
-        setMinimumSize(new Dimension(550, 460));
         setLocationRelativeTo(App.launcher.getParent());
         setLayout(new BorderLayout());
         setResizable(true);